Add files via upload
The provided code performs a simple image processing task that converts a given image to a pencil sketch effect. The script reads an image file named "image1.jpg" using the OpenCV cv2.imread() function and displays it in a window using the cv2.imshow() function. Then it converts the color image to grayscale using the cv2.cvtColor() function and inverts the grayscale image using the cv2.bitwise_not() function.